create handle error utils function
