+++ src/actions/course_tables.js
@@ -1,5 +1,6 @@
 import { UPDATE_COURSE_TABLE } from "constants/action-types";
 import instance from "api/axios";
+import handleAPIError from "utils/handleAPIError";
 
 const createCourseTable = (course_table_id, course_table_name, user_id, semester) => async (dispatch) => {
   try {
@@ -9,38 +10,7 @@ const createCourseTable = (course_table_id, course_table_name, user_id, semester
     dispatch({ type: UPDATE_COURSE_TABLE, payload: course_table });
     return course_table;
   } catch (error) {
-    // console.log(Error("Error in createCourseTable: "+error));
-
-    if (error.response) {
-      // server did response, used for handle custom error msg
-      const error_obj = {
-        status_code: error.response.status,
-        backend_msg: error.response.data.message,
-        error_info: error.message,
-        error_detail: Error(error).stack,
-      };
-      throw error_obj;
-    } else if (error.request) {
-      // The request was made but no response was received (server is downed)
-      const status = 521; // Server is down
-      const error_obj = {
-        status_code: status,
-        backend_msg: "no",
-        error_info: error.message,
-        error_detail: Error(error).stack,
-      };
-      throw error_obj;
-    } else {
-      // Something happened in setting up the request that triggered an Error
-      const status = 400; // Bad request
-      const error_obj = {
-        status_code: status,
-        backend_msg: "no",
-        error_info: error.message,
-        error_detail: Error(error).stack,
-      };
-      throw error_obj;
-    }
+    throw handleAPIError(error);
   }
 };
 
@@ -52,8 +22,6 @@ const fetchCourseTable = (course_table_id) => async (dispatch) => {
     dispatch({ type: UPDATE_COURSE_TABLE, payload: course_table });
     return course_table;
   } catch (error) {
-    // console.log(Error("Error in fetchCourseTable: "+error));
-
     if (error.response) {
       if (error.response.status === 403 || error.response.status === 404) {
         // expired course_table
@@ -62,36 +30,7 @@ const fetchCourseTable = (course_table_id) => async (dispatch) => {
         return null;
       }
     } else {
-      if (error.response) {
-        // server did response, used for handle custom error msg
-        const error_obj = {
-          status_code: error.response.status,
-          backend_msg: error.response.data.message,
-          error_info: error.message,
-          error_detail: Error(error).stack,
-        };
-        throw error_obj;
-      } else if (error.request) {
-        // The request was made but no response was received (server is downed)
-        const status = 521; // Server is down
-        const error_obj = {
-          status_code: status,
-          backend_msg: "no",
-          error_info: error.message,
-          error_detail: Error(error).stack,
-        };
-        throw error_obj;
-      } else {
-        // Something happened in setting up the request that triggered an Error
-        const status = 400; // Bad request
-        const error_obj = {
-          status_code: status,
-          backend_msg: "no",
-          error_info: error.message,
-          error_detail: Error(error).stack,
-        };
-        throw error_obj;
-      }
+      throw handleAPIError(error);
     }
   }
 };
@@ -111,8 +50,6 @@ const patchCourseTable = (course_table_id, course_table_name, user_id, expire_ts
     dispatch({ type: UPDATE_COURSE_TABLE, payload: course_table });
     return course_table;
   } catch (error) {
-    // console.log(Error("Error in patchCourseTable: "+error));
-
     // need to let frontend handle error, so change to return null
     if (error.response) {
       if (error.response.status === 403 && error.response.data.message === "Course table is expired") {
@@ -122,36 +59,7 @@ const patchCourseTable = (course_table_id, course_table_name, user_id, expire_ts
         return null;
       }
     } else {
-      if (error.response) {
-        // server did response, used for handle custom error msg
-        const error_obj = {
-          status_code: error.response.status,
-          backend_msg: error.response.data.message,
-          error_info: error.message,
-          error_detail: Error(error).stack,
-        };
-        throw error_obj;
-      } else if (error.request) {
-        // The request was made but no response was received (server is downed)
-        const status = 521; // Server is down
-        const error_obj = {
-          status_code: status,
-          backend_msg: "no",
-          error_info: error.message,
-          error_detail: Error(error).stack,
-        };
-        throw error_obj;
-      } else {
-        // Something happened in setting up the request that triggered an Error
-        const status = 400; // Bad request
-        const error_obj = {
-          status_code: status,
-          backend_msg: "no",
-          error_info: error.message,
-          error_detail: Error(error).stack,
-        };
-        throw error_obj;
-      }
+      throw handleAPIError(error);
     }
   }
 };
